Add bracket matcher and parse pwd
to support switching between shh (a customized ghci) and zsh.

Both shh and zsh prompts:
+ arent a constant because they reflect current directory (moreover zsh with zprezto
  checks current directory contents by executing 'ls' and then arising output,
  so promt easily might be 0.5 kb in size)
+ output many ansi-escape sequences

So our strategy is to detect prompt by several special characters at the
beggining and at the end of the promt; For example for a prompt like
$ [/tmp/test] ~ we might use '$ [' as a prefix and '] ~' as a suffix of
a prompt.

Matcher.Bracket is a hungry matcher which sequentelly consumes a stream
of characters; first detects prefix, then suffix and considers
everything in between as a content of a match.

Matcher.Unboxed defines a common OOP-style interface for Matcher.Seq and
Matcher.Bracket. Matcher.ByteString and Matcher.Text are specialized
versions to work with sequences of Word8 or Char characters.

For zsh we can just read /proc/<pid>/cwd

ghci (and hence shh) uses current working directory to search for
modules; because of that :cd causes it to unload loaded modules and it's
not practical to use :cd while using ghci as an interactive shell;
because of that ghci (or shh) keeps track of what setCurrentDirectory does
and then changes directory before executing a shell-like command and changes
it back afterwards. That way ghci doesn't need to unload modules and we
can do shell-like scripting in ghci with cd. Unfortunately because of
that most of the time /proc/<pid>/cwd reflects ghci working directory,
not a working directory used to execute shell-like commands. So we have
to run "pwd" and parse it's output to get relevant cwd.

shh wraps output of each command (like pwd) in a few ansi escape sequences,
so that we need to strip them. Likely there is a library to do that.