Optimize Matcher.Seq.ByteString.match_str
by inlining definition of matchStep manually. And also by
adding a special case for skipping input untill the first character
of a pattern.

The problem with matchStep is that it's a self-recursive function
so it can't be inlined. In the same time semantically it should
be a loop body iterating over elements of an input sequence.

For now we can manually inline definition and observe 3times
performance improvement.

We further optimize by writing a special case for skipping the input
untill the first character of a pattern. We think that the reason
why it improves the performance is because ByteString needs to check
boundaries and do something like withForeignPtr access each element.
BS.dropWhile does boundaries check and withForeignPtr only
once and then just iterate over elements. It should compile into
a faster loop.