Add matchStr as a method of SomeMatcher
The idea of having a matcher with a single function interface
matcherStep didn't work well. For us it's important to call
a fast ByteString.break (== c) otherwise a matcher is slow.

Now we are facing a problem that matchStr implementation is exactly the
same for ByteStrings and for Text except that we use functions from
either of two packages.