Copy output of previous shell cmd after switch
Change shell detection stratage for zsh. It turned out that detecting
prompt boundaries in case of zsh is quite a compex task becase:
in out test case (zsh with zprezto and cloud prompt theme with
interactive command autocompletion from history) autocompletion
constantly emits \b or \r and redraws a prompt together with
partually inputted command and proposed completion.

Instead of trying to detect the whole sequnce which corresponds to a prompt,
detect only last few bytes and then strip the first and the last lines of captured
output.