Fix resizing of tui puppets
Problem: a tui app started from a puppet shell doesn't redraw it's
tui if terminal was resized. But exiting/enterin an app draws
interface properly.
Cause: WINCH wasn't delivered to children of a puppet shell
Fix: deliver WINCH to a process group; AFAIU spawned children
become members of the same process group as a parent, so it
seem to work.