Fix jobs control
Problem: puppet bash complains that it cannot do jobs control
Cause:

Children processes inherit controlling terminal of a parent. This can be
changed if a process:
+ starts a new session (at this point it will have no controlling
  terminal)
+ acquires a "free"(not acquired by the others) controlling terminal

Fix:

The session should be started by a process itself. Since we cannot ask
an arbitrary program to do so, we implement a wrapper to:
+ start a new session
+ acquire a terminal
+ exec a new process

An additional executable will complicate packaging a bit.
Alternative solution would be to modify rts which seem to be out of
reach.