Remove polymorphism from Matchers
Making Matchers polymorphic was an experiment to answer the question
"can we make a polymorphic implementation and then specialize it to be
as fast as monomorphic one". It is an artificial problem which was
brought up for educational purposes which we are no longer interested to
support.

Time profiling shows that before this change ListLike.break takes a lot
of time. That means that optimized version of ByteString.break (==c) wasn't
used which makes performance of out Matcher more than 100 times worse.

The exact reason why the optimizer didn't specialize the calls is an
interesting question. Most likely this happens because of using a
SomeValue pattern with existential type. But we didn't dig deep enough
to find the exact root cause and how to fix it.