Optimize Pipe and add toEv
Turn Pipe into a zipper which represents a sequence of Programs and
focuses a program from which we will read output in the next step.

toEv wraps a program into BuffInput node which can buffer one input
value. That way stepUnsafe can first read all outputs from a program and
then feed it with buffered input. This is useful in initialization code
to store ProgramEv but postpone processing of potential output until the
main loop.