Fix copying of last output after switch
disable automatic copying after switch.

Problem: prevCmdOut is stored per puppet. Just after switching a "true"
previous output is stored in a background puppet; the current puppet
stores older output from the time when it was in a foreground.

Fix: store prevCmdOut in a MuxState; if syncCwd is not working, remember
output from a current puppet; whyle syncCwd is working just keep an
older output.