Rework PuppetSt structure
Before we were always keeping a PuppetSt for each puppet regardless of
weather puppet's process is running or not. The only place where it's
useful is keeping PuppetMode or a current puppet cause it reflects the
terminal state.

After this change we keep PuppetSt only is puppet's process is running.
This change is helpful for multiple puppets support.