Fix isEv
Remove default case from pattern match to be notified about
non-exhaustive pattern match when adding new Program constructions.

Change implementation of isEv for Pipe. Before we would go trough list of
non evaluated programs in a pipe to check if they are actually
evaluated. This is unnecessary because the contract for Pipe states that
if it was completely evaluated then list of not evaluated is empty.

Indeed, due to implementation details of interpreter we might be in a
situation when calling stepOutput to evaluate an unevaluated program will
not change it which means it was already evaluated. Hoverer correct usage
of Program is to call stepOut until we evaluate program. The new
implementation of isEv is coherent with that behaviour of stepUnsafe.